6.24 배틀시스템 최종본: BattleSystem 보스 생성 로직 구현, 3종의 보스 중 택 1로 출력, 1~ 1.5배 체력 및 공격력 계수 반영 Monster : h와 cpp 파일, SetHealth 의 randomPerHP, Atk는 int 파일명으로 인해 소수점 이하의 배수를 선정할 수 없어 float 파일명으로 수정 세부적인 레벨 디자인 필요 시 수정 가능

+++ TextRPG_pjt/TextRPG_pjt/Monster.cpp
@@ -51,7 +51,7 @@ void Monster::SetHealth(int playerLevel, int randomPer)
 }
 
 //레벨당 체력 공격력 조정 (플레이어 레벨 + 적용 퍼센트Hp + 적용 퍼센트 Atk)
-void Monster::SetHealth(int playerLevel, int randomPerHp, int randomPerAtk)
+void Monster::SetHealth(int playerLevel, float  randomPerHp, float  randomPerAtk)
 {
 	health_ = (health_ * (playerLevel)) * randomPerHp;
 	attack_ = (attack_ * (playerLevel)) * randomPerAtk;
